Added sample data

+++ src/main/java/analysis/TrendAnalysis.java
@@ -1,21 +1,20 @@
 package analysis;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
  * Represents trend analysis results.
  */
 public class TrendAnalysis {
-    private List<BigDecimal> weeklySpending;
+    private List<Double> weeklySpending;
     private FinancialAnalysis.TrendDirection trendDirection;
     
-    public TrendAnalysis(List<BigDecimal> weeklySpending, FinancialAnalysis.TrendDirection trendDirection) {
+    public TrendAnalysis(List<Double> weeklySpending, FinancialAnalysis.TrendDirection trendDirection) {
         this.weeklySpending = weeklySpending;
         this.trendDirection = trendDirection;
     }
     
-    public List<BigDecimal> getWeeklySpending() { return weeklySpending; }
+    public List<Double> getWeeklySpending() { return weeklySpending; }
     public FinancialAnalysis.TrendDirection getTrendDirection() { return trendDirection; }
     
     @Override
